+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "client",
+  "name": "client-mern",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
