+++ client/src/index.js
@@ -1,11 +1,11 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './components/App/App';
-import reportWebVitals from './reportWebVitals';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import App from "./components/App/App";
+import reportWebVitals from "./reportWebVitals";
 import { Provider } from "react-redux";
 
-import store from './store'
+import store from "./store";
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
@@ -14,6 +14,5 @@ ReactDOM.render(
   </React.StrictMode>,
   document.getElementById("root")
 );
-
 
 reportWebVitals();
