+++ package.json
@@ -1,9 +1,10 @@
 {
-  "name": "mern-n",
+  "name": "n-mern",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
+    "test": "mocha",
     "server": "cd server && npm start ",
     "client": "cd client && npm install && npm install --only=dev --no-shrinkwrap && npm run build",
     "start": "concurrently \"npm run server\" \"npm run client\"",
@@ -23,5 +24,9 @@
   "dependencies": {
     "concurrently": "^5.3.0",
     "react-router-dom": "^5.2.0"
+  },
+  "devDependencies": {
+    "chai": "^4.3.0",
+    "mocha": "^8.2.1"
   }
 }
