+++ server/package.json
@@ -1,10 +1,11 @@
 {
-  "name": "app1",
+  "name": "server-mern",
   "version": "0.0.0",
   "private": true,
   "scripts": {
     "start": "node ./bin/www",
-    "dev": "nodemon ./bin/www"
+    "dev": "nodemon ./bin/www",
+    "test": "mocha"
   },
   "dependencies": {
     "bcrypt": "^5.0.0",
@@ -32,5 +33,10 @@
     "passport-local": "^1.0.0",
     "sqlite3": "^5.0.0",
     "validator": "^13.5.2"
+  },
+  "devDependencies": {
+    "chai": "^4.3.0",
+    "chai-http": "^4.3.0",
+    "mocha": "^8.2.1"
   }
 }
